Refactor langchain_text_splitters: rename nltk.py to nltk_spl.py, move NLTKTextSplitter to persian.py, update nltk and setuptools versions in poetry.lock, and modify pyproject.toml
