+++ libs/text-splitters/langchain_text_splitters/__init__.py
@@ -44,7 +44,7 @@ from langchain_text_splitters.markdown import (
     MarkdownHeaderTextSplitter,
     MarkdownTextSplitter,
 )
-from langchain_text_splitters.nltk import NLTKTextSplitter
+from langchain_text_splitters.nltk_spl import NLTKTextSplitter
 from langchain_text_splitters.python import PythonCodeTextSplitter
 from langchain_text_splitters.sentence_transformers import (
     SentenceTransformersTokenTextSplitter,
